POST routes for genes, genomes ans signal-genes
