+++ mist-api/src/routes/genes/post.js
@@ -3,8 +3,8 @@
 const getPostHelpers = require('./get-post-helpers')
 
 module.exports = function(app, middlewares, routeMiddlewares) {
-  const isPOST = false
-  return getPostHelpers.genomeFinder(app, middlewares, isPOST)
+  const isPOST = true
+  return getPostHelpers.geneFinder(app, middlewares, isPOST)
 }
 
 module.exports.docs = getPostHelpers.docs
