+++ mist-api/src/routes/genomes/$accession/^exists/signal-genes/get-post-helpers.js
@@ -1,0 +1,121 @@
+'use strict'
+
+// Vendor
+const bodyParser = require('body-parser')
+const _ = require('lodash')
+const { Op } = require('sequelize')
+
+// Local
+const {
+	scrub,
+	splitAndScrubString,
+} = require('core-lib/util')
+
+exports.signalGeneFinder = function(app, middlewares, isPOST = false) {
+    const models = app.get('models')
+	const helper = app.get('lib').RouteHelper.for(models.SignalGene)
+	const signalTransductionVersion = app.get('config').signalTransduction.version
+    const sequelize = app.get('sequelize')
+
+    const specificMiddlewares = []
+    if (isPOST) {
+        specificMiddlewares.push(bodyParser.urlencoded({extended: false}))
+    }
+
+    specificMiddlewares.push(
+        middlewares.parseCriteriaForMany(models.SignalGene, {
+            accessibleModels: [
+                      models.Component,
+                      models.Gene,
+                      models.Aseq,
+            ],
+            maxPage: null,
+            permittedOrderFields: '*',
+            permittedWhereFields: [
+              'id',
+              'gene_id',
+                      'component_id',
+                      'ranks'
+            ],
+          }, isPOST),
+    )
+
+    specificMiddlewares.push(
+        (req, res, next) => {
+			res.locals.criteria.include.push({
+				model: models.Component,
+				attributes: ['name', 'version', 'definition'],
+				where: {
+					genome_id: res.locals.genome.id
+				},
+				required: true
+			})
+			const userInput = isPOST ? req.body : req.query
+			const ranks = splitAndScrubString(userInput['where.ranks'], ',')
+			if (ranks.length) {
+				// Perform a contains query so that we can accomodate partial array matches.
+				// For example, finding all tcp proteins regardless of their second rank would
+				// not work with an exact match.
+				res.locals.criteria.where.ranks = {
+					[Op.contains]: ranks,
+				}
+			}
+
+			// Provide for searching by function / kind
+			const kind = scrub(userInput.kind)
+			if (!kind) {
+				next()
+				return
+			}
+
+			const func = scrub(userInput.function)
+			sequelize.transaction()
+			.then((transaction) => {
+				res.locals.criteria.transaction = transaction
+				return models.SignalDomain.getSignalDomainNames(signalTransductionVersion, kind, func, transaction)
+			})
+			.then((namesByKindAndFunction) => {
+				if (namesByKindAndFunction.length) {
+					// TODO: refactor this into a helper method that may be easily reused
+					const escapedArrayList = namesByKindAndFunction
+						.map((name) => sequelize.escape(name))
+						.join(',')
+					const queryInterface = sequelize.getQueryInterface()
+					const tableName = queryInterface.quoteTable(models.SignalGene.name)
+					const field = queryInterface.quoteIdentifier('counts')
+					_.set(res.locals.criteria, ['where', Op.and], sequelize.literal(`${tableName}.${field} ?| array[${escapedArrayList}]`))
+					next()
+					return
+				}
+
+				// User provided kind and function, but there are no corresponding rows in the database
+				// matching those query values. Thus, return no results.
+				res.json([])
+			})
+			.catch(next)
+        }
+    )
+
+    specificMiddlewares.push(helper.findManyHandler())
+
+    return specificMiddlewares
+}
+
+module.exports.docs = function(modelExamples) {
+	return {
+		name: 'Fetch Member Signal Genes',
+		description: 'Returns an array of <a href="#signal-gene-model">Signal Genes</a> that belong to the genome identified by ${accession}.',
+		example: {
+			request: {
+				parameters: {
+					accession: modelExamples.Genome.version
+				}
+			},
+			response: {
+				body: [
+					modelExamples.SignalGene
+				]
+			}
+		}
+	}
+}
